move exception to "Exceptions" directory to match structure of other packages
